feat: Transforming Dialog to MDX
